# PERF: remove useless overrides