- [x] closes #21522
- [x] tests passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [x] whatsnew entry

Asv run:

```
       before           after         ratio
     [9e982e18]       [705f0e3b]
-         243±7ms          186±4ms     0.77  multiindex_object.GetLoc.time_large_get_loc_warm
-       220±0.9ms          151±3ms     0.69  multiindex_object.GetLoc.time_large_get_loc
-         173±1ms          101±2ms     0.59  multiindex_object.Integer.time_get_indexer

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```

Notice that as of now the same _cannot_ be done for ``.is_monotonic_increasing`` and friends, because the sortedness of the ``MultiIndex`` corresponds to the sortedness of the underlying int index only if levels are themselves sorted.